cleanup epoch marker
